requirements.txt
alembic==1.12.1
blinker==1.7.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
flask-sock==0.7.0
h11==0.14.0
identify==2.5.31
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
ruff==0.1.5
setuptools==68.2.2
simple-websocket==1.0.0
SQLAlchemy==2.0.23
Turbo-Flask==0.8.4
typing_extensions==4.8.0
virtualenv==20.24.6
Werkzeug==3.0.1
wsproto==1.2.0
